pyproject.toml
[project]
name = "lecture-feedback"
version = "0.1.0"
description = "Streamlit application for tracking lecture audience feedback"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "streamlit>=1.49.1",
    "streamlit-autorefresh>=1.0.1",
]

[dependency-groups]
dev = [
    "pre-commit>=4.3.0",
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.13.3",
    "mypy>=1.13.0",
    "coverage>=7.12.0",
    "pytest-bdd>=8.1.0",
    "pandas-stubs>=2.3.3.251219",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/lecture_feedback"]

[tool.ruff]
# Set line length to match common Python standards
line-length = 88


[tool.ruff.lint]
# Enable comprehensive linting rules
select = [ "ALL" ]
ignore = [
    "D", # Docs enforcement is annoying, we write expressive code
]

[tool.ruff.lint.isort]
known-first-party = ["lecture_feedback"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = [
    "S101",  # Ignore use of assert detected in tests
    "PLR2004", # Allow magic numbers in tests
]


[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
show_error_codes = true
mypy_path = "src"

[[tool.mypy.overrides]]
module = [
    "streamlit.*",
    "streamlit_autorefresh.*",
    "plotly.*",
]
ignore_missing_imports = true
